+++ src/constants/actionTypes.js
@@ -13,6 +13,12 @@ export const LOGIN_REQUEST = 'LOGIN_REQUEST';
 export const LOGIN_PENDING = 'LOGIN_PENDING';
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 export const LOGIN_FAIL = 'LOGIN_FAIL';
+
+export const LOGOUT_REQUEST = 'LOGOUT_REQUEST';
+export const LOGOUT_PENDING = 'LOGOUT_PENDING';
+export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
+export const LOGOUT_FAIL = 'LOGOUT_FAIL';
+
 export const RECEIVE_USERNAME = 'RECEIVE_USERNAME';
 
 export const SET_FILTER = 'SET_FILTER';