+++ src/reducers/login.js
@@ -1,4 +1,11 @@
-import { LOGIN_PENDING, LOGIN_SUCCESS, LOGIN_FAIL } from '../constants/actionTypes';
+import {
+  LOGIN_PENDING,
+  LOGIN_SUCCESS,
+  LOGIN_FAIL,
+  LOGOUT_PENDING,
+  LOGOUT_SUCCESS,
+  LOGOUT_FAIL,
+} from '../constants/actionTypes';
 
 const login = (
   state = {
@@ -12,8 +19,14 @@ const login = (
       return { isLoggingIn: true, isLoggedIn: false };
     case LOGIN_SUCCESS:
       return { isLoggingIn: false, isLoggedIn: true };
-    case LOGIN_FAIL:
+    case LOGOUT_PENDING:
+      return { isLoggingIn: true, isLoggedIn: true };
+    case LOGOUT_SUCCESS:
       return { isLoggingIn: false, isLoggedIn: false };
+    case LOGIN_FAIL:
+      return { ...state, isLoggingIn: false };
+    case LOGOUT_FAIL:
+      return { ...state, isLoggingIn: false };
     default:
       return state;
   }